Music has been implemented
Music has been sourced from freesound.org and the beginning of the implementation has begun. At the moment I am running into issues with having the music have an effect on the final code. The play/pause button has been put into the code and it grows and shrinks with the size of the window.