Music has been added to the code.
The beginnings of adding music and having it be influence the final code has begun. The music has been sourced from freesound.org. The play/pause button has also been added and it grows and shrinks with the size of the window. At the moment I am running into issues with the music having an effect on the code but I am starting to make progress.